requirements.txt
Flask
ultralytics
Pillow
opencv-python-headless
gdown 